refactor: SearchToilet 페이지 전용 폴더 생성 및 파일 이동
